+++ src/data/languages.ts
@@ -2,3 +2,5 @@ export const LANGUAGES = [
   { code: "en", label: "English" },
   { code: "he", label: "עברית" },  
 ];
+
+export const RTL_LANGS = ["he", "ar", "fa", "ur"];
